requirements.txt
torch>=1.10
torchvision>=0.11
numpy>=1.19
pillow>=8.0
scikit-learn>=0.24
